Misc HW patches

- Remove versioned machines released in QEMU 3.1 and 4.0
- Enable X.Org text rendering on ATI VGA model
- Various memory leaks / overflows fixed
- Build using -Wformat-overflow=2 CPPFLAG

# -----BEGIN PGP SIGNATURE-----
#
# iQIzBAABCAAdFiEE+qvnXhKRciHc/Wuy4+MsLN6twN4FAmmu7BAACgkQ4+MsLN6t
# wN4P2RAA2m5XpiwqjyUaaidGQTZi1qNqj3DOl1nOh3XpFCzD4BGUtbJp+06i59iL
# Q8yxFr7jtsNDiRdWC2ow8rWp3aQhqzvsn1hpNFWZoVHWOBsPnHpOmgFINlw3nZA2
# d/boV8aKE/VXhLY8plL3T4oy9rUSiwd7S4rubYjVIr2BVOQaiMKEZvLYUS4ErP5Q
# rhhgj08Z6NhB6fE+kxmi/R1+C8i8mYnCtK2vO9kXoyfdiwFNBpfmaZ6xJ0eSI/v9
# 4GVkeEtTkBsvRjbo/S5OQIjUKYsB8y00UnaiO+VRPxVDHqks5eP2WtrZOjOKTs+C
# EaNkhqocMowGRt23JqM0XVaYktjzsXIYTGnRI33hmwL1qbyYK91wzxt5JqwYlpHo
# EGn6YqV4to/1QLSzPr25cETf089ZJ6gqoIPP/I8dwj4Ups+nDaAvRvOKAOTTnyez
# F8zZ+AYe6jhZKKN0iTwoaBUCFt1VgtJ6gSz/GChoALTrEcOGEzBxkfAbzTq9ir8/
# lmO62Rmdrzc8KZiclOyCQM1kOzNH4Jdzy3jEN9cyIQRhf+KVz1uQEqyR4Lb8xmHy
# S9Hts6CTP8yPXh/kL0xAlaY1IDRrkFwuTS52+XjwNyKTQFsOJX6VE2fJgA8ut7Jq
# gNc3sq0G1aCMRqB/1XHvawfSK1o6Olyk1P1i/FUAFzjIxGX0PMw=
# =zL9Q
# -----END PGP SIGNATURE-----
# gpg: Signature made Mon Mar  9 15:49:36 2026 GMT
# gpg:                using RSA key FAABE75E12917221DCFD6BB2E3E32C2CDEADC0DE
# gpg: Good signature from "Philippe Mathieu-Daudé (F4BUG) <f4bug@amsat.org>" [full]
# Primary key fingerprint: FAAB E75E 1291 7221 DCFD  6BB2 E3E3 2C2C DEAD C0DE

* tag 'hw-misc-20260309' of https://github.com/philmd/qemu: (49 commits)
  hw/misc/riscv_cpc: Free CPU array in instance_finalize
  hw/misc/sifive_e_aon: Don't leak timer
  hw/i3c/dw-i3c: Fix memory leaks in error logging paths
  ati-vga: Do not access pixel outside the screen
  ati-vga: Simplify hw cursor drawing
  ati-vga: Implement HOST_DATA flush to VRAM
  ati-vga: Implement HOST_DATA register writes
  ati-vga: Implement scissor rectangle clipping for 2D operations
  ati-vga: Remove ATIVGAState param from ati_2d_do_blt
  ati-vga: Split ati_2d_do_blt from ati_2d_blt
  ati-vga: Extract setup_2d_blt_ctx from ati_2d_blt
  ati-vga: Introduce ATI2DCtx struct for 2D blit context
  ati-vga: Use local variables for register values in ati_2d_blt
  ati-vga: Remove src and dst stride mutation in ati_2d_blt
  hw: Make qdev_get_printable_name() consistently return freeable string
  hw/qdev: Document qdev_get_dev_path()
  hw/timer/slavio_timer: Free resources allocated in instance_init
  hw/sparc/sun4m: Use qdev GPIOs rather than qemu_allocate_irqs()
  hw/sparc/sun4m: Don't set up dummy cpu interrupts
  hw/net/npcm_gmac: Catch accesses off the end of the register array
  ...

+++ net/passt.c
@@ -102,7 +102,9 @@ static void net_passt_cleanup(NetClientState *nc)
     }
 #endif
 
-    kill(s->pid, SIGTERM);
+    if (s->pid > 0) {
+        kill(s->pid, SIGTERM);
+    }
     if (g_remove(s->pidfile) != 0) {
         warn_report("Failed to remove passt pidfile %s: %s",
                     s->pidfile, strerror(errno));
@@ -268,8 +270,17 @@ static int net_passt_start_daemon(NetPasstState *s, int sock, Error **errp)
         return -1;
     }
 
-    if (g_subprocess_get_if_exited(daemon) &&
-        g_subprocess_get_exit_status(daemon)) {
+    if (g_subprocess_get_if_exited(daemon)) {
+        gint status = g_subprocess_get_exit_status(daemon);
+        if (status) {
+            error_setg(errp, "Passt exited with code %d", status);
+            return -1;
+        }
+    }
+
+    if (g_subprocess_get_if_signaled(daemon)) {
+        error_setg(errp, "Passt killed with signal %d",
+                   g_subprocess_get_term_sig(daemon));
         return -1;
     }
